BE-feat-참가한 크루잉 모아보기 기능/ 크루잉 참가 memberId 검증/ 리프레쉬 토큰 기능 구조 변경

+++ server/mainProject/src/main/java/com/codestates/mainProject/crewing/service/CrewingService.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -26,6 +28,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static org.springframework.data.redis.connection.RedisStringCommands.SetOption.ifPresent;
 
 @Service
 @Transactional
@@ -44,7 +48,7 @@ public class CrewingService {
     }
 
     /** 게시글 생성 */
-    public Crewing createCrewing(Crewing crewing) {
+    public Crewing createCrewing(@RequestBody Crewing crewing) {
         /** JWT토큰정보를 이용한 사용자 인증 */
         String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
         Optional<Member> verifiedMember = memberRepository.findByEmail(principal);
@@ -52,11 +56,14 @@ public class CrewingService {
         Member member = verifiedMember.
                 orElseThrow(() -> new BusinessLogicException(ExceptionCode.NO_PERMISSION));
 
+        if (crewing.getMember().getMemberId() != member.getMemberId()) {
+            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);
+        }
+
         CrewingMembers crewingMember = new CrewingMembers();
         crewingMember.setMember(member);
         crewingMember.setCrewing(crewing);
         crewing.setCurrentPeople(0); // 초기 값은 0으로 설정
-        crewing.setCompleted(false); // 초기 값은 false로 설정
 
         return crewingRepository.save(crewing);
     }
@@ -65,9 +72,9 @@ public class CrewingService {
     public Crewing updateCrewing(Crewing crewing) {
         Crewing findCrewing = findVerifiedCrewing(crewing.getCrewingId());
         /** JWT토큰정보를 이용한 사용자 인증 */
-        /*String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
-        if (!findPost.getMember().getEmail().equals(principal))
-            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);*/
+        String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        if (!findCrewing.getMember().getEmail().equals(principal))
+            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);
 
         Optional.ofNullable(crewing.getTitle())
                 .ifPresent(title -> findCrewing.setTitle(title));
@@ -89,32 +96,48 @@ public class CrewingService {
                 .ifPresent(isCompleted -> findCrewing.setCompleted(isCompleted));
         return crewingRepository.save(findCrewing);
     }
-    public void canApply(long crewingId, CrewingDto.applyDto apply){
+
+    /** 크루잉 참여신청 예외 처리 */
+    public String canApply(long crewingId, CrewingDto.applyDto apply){
+        /** Crewing 및 Member 확인 */
         Crewing crewing = crewingRepository.findByCrewingId(crewingId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.CREWING_NOT_FOUND));
         Member member = memberRepository.findById(apply.getMemberId())
                 .orElseThrow(()-> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+        /** 본인이 작성한 게시글에는 신청할 수 없음, Code:403 */
         if(crewing.getMember().getMemberId().equals(apply.getMemberId())){
-            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);
+            throw new BusinessLogicException(ExceptionCode.CREWING_OWN_APPLY);
         }
+        /** 모집마감일이 지났다면 신청할 수 없음, Code:403 */
         LocalDateTime now = LocalDateTime.now();
         LocalDateTime deadline = LocalDateTime.parse(crewing.getDeadLine(), DateTimeFormatter.ISO_DATE_TIME);
         if(now.isAfter(deadline)){
-            throw new BusinessLogicException(ExceptionCode.CREWING_IS_CLOSED);
+            crewing.setCompleted(true);
+            throw new BusinessLogicException(ExceptionCode.CREWING_DEADLINE);
         }
+        /** 현재인원이 모집인원보다 같거나 크다면 신청할 수 없음, Code:403 */
         int currentPeople = crewingMembersRepository.countByCrewing(crewing);
-
-        if(crewing.getMaxPeople() >= currentPeople){
-            applyCrewing(crewingId, apply.getMemberId(),crewing, member, currentPeople);
-        } else {
+        if(crewing.getMaxPeople() < currentPeople){
             throw new BusinessLogicException(ExceptionCode.CREWING_IS_MAX);
         }
+
+       return applyCrewing(crewingId, apply.getMemberId(),crewing, member, currentPeople);
     }
-    public void applyCrewing(long crewingId, long memberId, Crewing crewing, Member member, int currentPeople){
+
+    /** 크루잉 참여신청 */
+    public String applyCrewing(long crewingId, long memberId, Crewing crewing, Member member, int currentPeople){
+        String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        if (!memberRepository.findById(memberId).orElseThrow(()->new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND)).getEmail().equals(principal))
+            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);
         CrewingMembers existApply = crewingMembersRepository.findByMemberAndCrewing(member, crewing);
         if(existApply!=null){
             crewingMembersRepository.delete(existApply);
             currentPeople--;
+            crewing.setCurrentPeople(currentPeople);
+            if(crewing.isCompleted() && crewing.getMaxPeople()>currentPeople){
+                crewing.setCompleted(false);
+            }
+            return "Crewing application canceled.";
         } else if(!crewing.isCompleted()){
             CrewingMembers.CrewingMemberId crewingMemberId = new CrewingMembers.CrewingMemberId();
             crewingMemberId.setCrewingId(crewingId);
@@ -126,33 +149,24 @@ public class CrewingService {
             crewingMembersRepository.save(crewingMembers);
             currentPeople++;
         } else {
-            throw new BusinessLogicException(ExceptionCode.CREWING_IS_MAX);
+            throw new BusinessLogicException(ExceptionCode.CREWING_IS_CLOSED);
         }
+        /** 이미 모집이 마감되었다면 신청할 수 없음, Code:403 -> 이미 신청했던 회원은 신청취소되는 로직 다음에 와야함 */
+
         crewing.setCurrentPeople(currentPeople);
+
         if(crewing.getMaxPeople()==currentPeople){
             crewing.setCompleted(true);
-        } else if(crewing.isCompleted() && crewing.getMaxPeople()>currentPeople){
-            crewing.setCompleted(false);
         }
+        return "You have successfully applied to the crewing";
     }
 
     /** 게시글 조회 */
     public CrewingDto.ResponseDto getCrewing(long crewingId) {
         Crewing findCrewing = findVerifiedCrewing(crewingId);
-        List<CrewingMembers> crewingMembers = crewingMembersRepository.findByCrewing(findCrewing);
-        List<Member> Members = crewingMembers.stream()
-                .map(crewingMember -> crewingMember.getMember())
-                .distinct()
-                .collect(Collectors.toList());
-        List<CrewingDto.Members> CrewingMember = new ArrayList<>();
-        for(Member member : Members){
-            CrewingDto.Members members = new CrewingDto.Members();
-            members.setUserName(member.getUserName());
-            members.setImageUrl(member.getImageUrl());
-            CrewingMember.add(members);
-        }
+
         CrewingDto.ResponseDto Response = crewingmapper.crewingToCrewingResponse(findCrewing);
-        Response.setMembers(CrewingMember);
+
         return Response;
     }
 
@@ -174,9 +188,9 @@ public class CrewingService {
     public void deleteCrewing(long crewingId) {
         Crewing findCrewing = findVerifiedCrewing(crewingId);
         /** JWT토큰정보를 이용한 사용자 인증 */
-        /*String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
-        if (!findPost.getMember().getEmail().equals(principal))
-            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);*/
+        String principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        if (!findCrewing.getMember().getEmail().equals(principal))
+            throw new BusinessLogicException(ExceptionCode.NO_PERMISSION);
         crewingRepository.deleteById(crewingId);
     }
 
@@ -193,6 +207,27 @@ public class CrewingService {
         List<CrewingDto.ResponseDto> responseDto = crewingmapper.crewingListToCrewingResponseList(listCrewing);
         return new MultiResponseDto(responseDto,pageInfo);
     }
+    public MultiResponseDto getMyApply(Member member, int page, int size, Long lastPostId){
+        Pageable pageRequest = PageRequest.of(page-1, size, Sort.by("id.crewingId").descending());
+        Page<CrewingMembers> findCrewings;
+        if(lastPostId == null){
+            findCrewings = crewingMembersRepository.findByMember(member,pageRequest);
+        } else {
+            findCrewings = crewingMembersRepository.findByMemberAndIdCrewingIdLessThan(member, lastPostId, pageRequest);
+        }
+        List<Long> findCrewingId = findCrewings.stream()
+                .map(findCrewing -> findCrewing.getId().getCrewingId())
+                .distinct()
+                .collect(Collectors.toList());
+        List<CrewingDto.ResponseDto> response = new ArrayList<CrewingDto.ResponseDto>();
+        for(long crewingId : findCrewingId){
+            Crewing crewing = crewingRepository.findByCrewingId(crewingId)
+                    .orElseThrow(()-> new BusinessLogicException(ExceptionCode.CREWING_NOT_FOUND));
+            response.add(crewingmapper.crewingToCrewingResponse(crewing));
+        }
+        PageInfo pageInfo = new PageInfo(page,findCrewings.getSize(),findCrewings.getTotalElements(),findCrewings.getTotalPages(), findCrewings.hasNext());
+        return new MultiResponseDto(response, pageInfo);
+    }
 
     /** 게시글 존재하는지 확인 */
     public Crewing findVerifiedCrewing(long crewingId) {
@@ -204,6 +239,25 @@ public class CrewingService {
 
     public Page<Crewing> getCrewingsByIdLessThan(Long lastCrewingId, Pageable pageable) {
         return crewingRepository.findByCrewingIdLessThan(lastCrewingId, pageable);
+    }
+
+    @Scheduled(cron = "0 00 00 * * ?") /** Seconds, Minutes, Hours, Day of month, Month, Day of week, Year(Option) */
+    public void updateCompletedStatus() {
+        List<Crewing> crewings = crewingRepository.findAll();
+
+        for (Crewing crewing : crewings) {
+            if (isDeadlineReached(crewing)) {
+                crewing.setCompleted(true);
+                crewingRepository.save(crewing);
+            }
+        }
+    }
+
+    private boolean isDeadlineReached(Crewing crewing) {
+        LocalDateTime deadline = LocalDateTime.parse(crewing.getDeadLine());
+        LocalDateTime now = LocalDateTime.now();
+
+        return now.isAfter(deadline) || now.isEqual(deadline);
     }
 
 }
